fix: ctc alignment (#34)

+++ Assets/Scripts/Tale/Action/DialogAction.cs
@@ -461,7 +461,7 @@ namespace TaleUtil
                         }
                     }
 
-                    screenToWorldUnit = Screen.width / Config.REFERENCE_WIDTH;
+                    screenToWorldUnit = ((float) Screen.width) / Config.REFERENCE_WIDTH;
 
                     if (avatar != null)
                     {
